// src/app/shared/models/models.ts
export interface apiResultFormat {
  data: [];
  totalData: number;
}
// ...
export interface BookingDetails {
  isSelected?: boolean;
  sNo?: number;
  id: string;
  hotelName: string;
  hotelLocation: string;
  room: string;
  guest: string;
  days: string;
  pricing: string;
  bookedOn: string;
  status: string
}